@@ -1,54 +1,54 @@
 {
-  "name": "tecnodesignc/user-module",
-  "type": "encore-module",
-  "description": "User module for EncoreCMS. Handles the authentication and authorisation as well as the user management.",
-  "keywords": [
-    "tecnodesign",
-    "user",
-    "authentication",
-    "authorisation"
-  ],
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Marcos Rativa",
-      "email": "info@tecnodesign.com",
-      "role": "Developer"
-    }
-  ],
-  "support": {
-    "email": "support@tecnodesign.com.co",
-    "issues": "https://github.com/tecnodesignc/Platform/issues",
-    "source": "https://github.com/tecnodesignc/User"
-  },
-  "require": {
-    "php": ">=7.1.3",
-    "composer/installers": "~1.0",
-    "ramsey/uuid": "^4.1"
-  },
-  "require-dev": {
-    "phpunit/phpunit": "~7.0",
-    "orchestra/testbench": "3.8.*",
-    "scrutinizer/ocular": "~1.1",
-    "fzaninotto/faker": "~1.5",
-    "mockery/mockery": "^0.9.4"
-  },
-  "autoload": {
-    "psr-4": {
-      "Modules\\User\\": ".",
-      "Modules\\": "Modules/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Modules\\User\\Tests\\": "Tests"
-    }
-  },
-  "extra": {
-    "branch-alias": {
-      "dev-master": "1.0.x-dev"
-    }
-  },
-  "minimum-stability": "dev",
-  "prefer-stable": true
+    "name": "tecnodesignc/user-module",
+    "type": "encore-module",
+    "description": "User module for EncoreCMS 2. Handles the authentication and authorisation as well as the user management.",
+    "keywords": [
+        "tecnodesign",
+        "user",
+        "authentication",
+        "authorisation"
+    ],
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Marcos Rativa",
+            "email": "info@tecnodesign.com",
+            "role": "Developer"
+        }
+    ],
+    "support": {
+        "email": "support@tecnodesign.com.co",
+        "issues": "https://github.com/tecnodesignc/encorecms/issues",
+        "source": "https://github.com/tecnodesignc/user-module"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.0.x-dev"
+        }
+    },
+    "require": {
+        "php": ">=8.0",
+        "composer/installers": "^2.1",
+        "ramsey/uuid": "^4.1"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^9.5",
+        "orchestra/testbench": "3.8.*",
+        "scrutinizer/ocular": "~1.1",
+        "fzaninotto/faker": "~1.5",
+        "mockery/mockery": "^1.4.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "Modules\\User\\": ".",
+            "Modules\\": "modules/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Modules\\User\\Tests\\": "Tests"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
