@@ -11,7 +11,7 @@ class RoleIsUpdating extends AbstractEntityHook implements EntityIsChanging
     /**
      * @var RoleInterface
      */
-    private $role;
+    private RoleInterface $role;
 
     public function __construct(RoleInterface $role, $attributes)
     {
@@ -22,7 +22,7 @@ class RoleIsUpdating extends AbstractEntityHook implements EntityIsChanging
     /**
      * @return RoleInterface
      */
-    public function getRole()
+    public function getRole(): RoleInterface
     {
         return $this->role;
     }
